[package_godot_projects.sh] Refactor to fix various error

+++ programs/package_godot_projects.sh
@@ -60,8 +60,6 @@ $output_directory	 -- Directory to output the clean Godot projects and optional 
 }
 
 # Parses command-line options using getopts
-# Outputs positional arguments to a temporary file `$temp_file`, see `main()`.
-#
 # Arguments:
 # $@ -- The arguments passed to the program
 parse_cli_arguments() {
@@ -107,40 +105,44 @@ parse_cli_arguments() {
 		esac
 	done
 	shift $((OPTIND - 1))
-	$dir_godot="$1"
-	$dir_dist="$2"
+	dir_godot="$1"
+	dir_dist="$2"
 }
 
 package_godot_projects() {
-	dir_export="$dir_godot/$out_file_name"
+	dir_export="$dir_dist/$out_file_name"
+
+	temp_file=$(mktemp)
+	find "$dir_godot" -not -path "./tutorial" -not -path "./dist" -type f -name project.godot -exec dirname {} \; >>$temp_file
+	dirs_count=$(wc -l <$temp_file)
+	printf "Found %s godot projects." "$dirs_count"
+	test $dirs_count -eq 0 && return 0
 
 	if test $is_dry_run -eq 0; then
-		test ! -d "$dir_export" && mkdir "$dir_export"
-		test ! -d "$dir_dist" -a $do_create_dirs -eq 1 && mkdir "$dir_dist"
+		test ! -d "$dir_export" && mkdir -p "$dir_export"
+		test ! -d "$dir_dist" && mkdir -p "$dir_dist"
 	fi
 
-	godot_project_folders=$(find "$dir_godot" -not -path "./tutorial" -not -path "./dist" -type f -name project.godot -exec dirname {} \;)
-	printf "Found %s godot projects." "$(wc -l "$godot_project_folders")"
-	for i in $godot_project_folders; do
-		printf "Copying project %s to %s." "$(basename "$i")" "$dir_export"
-		test $is_dry_run -eq 0 && cp -r "$i" "$dir_export"
-	done
+	while read dirpath; do
+		printf "Copying project %s to %s.\n" "$(basename "$dirpath")" "$dir_export"
+		test $is_dry_run -eq 0 && cp -r "$dirpath" "$dir_export"
+	done <"$temp_file"
 
 	echo "Removing .import directories..."
-	test $is_dry_run -eq 0 && rm -rf "$(find . -path dist -type d -name .import)"
+	test $is_dry_run -eq 0 && rm -rf $(find . -path dist -type d -name .import)
 	echo "Done."
 
 	if test $do_zip -eq 1; then
 		if test $is_dry_run -eq 0; then
 			archive_name="$out_file_name.zip"
-			zip -r "$archive_name" "$dir_export/*"
-			mv -v --backup --force "$archive_name" "$dir_dist"
+			zip -r "$archive_name" $(find "$dir_export" -maxdepth 1 -type d) >/dev/null
+			mv -v --force "$archive_name" "$dir_dist"
 			rm -rf "$dir_export"
 		fi
 		echo "Removing the $dir_export directory..."
 	else
 		echo "Moving directory $dir_export to $dir_dist."
-		test $is_dry_run -eq 0 && mv -r --backup --force "$dir_export" "$dir_dist"
+		test $is_dry_run -eq 0 && mv -rf "$dir_export" "$dir_dist"
 	fi
 	echo "Done."
 }
@@ -152,15 +154,8 @@ main() {
 	local dir_godot=""
 	local dir_dist=""
 	local do_zip=1
-	local do_create_dirs=0
 
 	parse_cli_arguments "$@"
-
-	if test ! -d "$dir_godot" -o ! -d "$dir_dist"; then
-		printf "%s: Both directories %s and %s must exist. Aborting." "$(format_bold Error)" "$dir_godot" "$dir_dist"
-		exit 1
-	fi
-
 	package_godot_projects
 }
